requirements.txt
fastapi
uvicorn
jinja2
httpx
feedparser
pydantic
apscheduler
python-dotenv
